+++ apps/editor/app/app.cpp
@@ -68,7 +68,7 @@ auto editor_app::on_file_drop(const std::vector<std::string>& paths) -> void
                     if (std::shared_ptr<sbk::engine::sound> createdSound =
                             project->create_database_object<sbk::engine::sound>())
                     {
-                        createdSound->set_database_name(newFilePath.filename().stem().string());
+                        createdSound->set_object_name(newFilePath.filename().stem().string());
                         createdSound->set_sound_name(
                             std::filesystem::relative(newFilePath, project->get_config().source_folder()).string());
                     }