+++ src/sound_bakery/core/object/object_tracker.h
@@ -4,7 +4,7 @@
 
 namespace sbk::core
 {
-    struct object_ptr_comparator
+    struct SB_CLASS object_ptr_comparator
     {
         bool operator()(const object* lhs, const object* rhs) const;
     };