+++ src/sound_bakery/editor/editor_defines.h
@@ -8,9 +8,9 @@ namespace sbk::editor
 
     enum class METADATA_KEY
     {
-        Payload,
+        payload,
         MinMax,
-        Readonly
+        readonly
     };
 
     inline std::string PayloadObject        = "OBJECT";