+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Drag & Dropping .wav files copies them to the project
 - Drag & Dropping .bakery files opens the project
 - The editor has a new "Create New Project" dialog window
+- The database_object class has a get_database_name and get_database_path function for generating unique object identifiers
 
 ### Removed
 
@@ -21,12 +22,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Editor opens in fullscreen upon start
 - Editor strings are shown in alphabetical order instead of random
 - The sound_bakery and sound_chef libraries are replaced in favour of xxx_shared and xxx_static libraries
+- Object names are moved from the database_object class to the object class
+- **BREAKING**: Public type names have changed and Sound Bakery will not be able to load old files
 
 ### Fixed
 
 - Fixed assert failing when a parent node is stopping a child
 - Fixed saving sounds when no encoded sound exists
 - Fixed the editor save icons displaying a question mark
+- Sound nodes without sounds don't trigger an assert and correctly stop
 
 ## [0.3.0]